+++ src/main/java/sample/Views/ResponseCode.java
@@ -1,4 +1,4 @@
-package sample;
+package sample.Views;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
