+++ src/main/java/sample/Views/UserInfo.java
@@ -1,7 +1,29 @@
 package sample.Views;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
 /**
  * Created by ksg on 17.03.17.
  */
-public class UserInfo {
+public final class UserInfo {
+    final String userMail;
+    final String userLogin;
+
+    @JsonCreator
+    public UserInfo(@JsonProperty("userMail") String userMail,
+                    @JsonProperty("userLogin") String userLogin) {
+        this.userLogin = userLogin;
+        this.userMail = userMail;
+    }
+
+    @SuppressWarnings("unused")
+    public String getUserLogin() {
+        return userLogin;
+    }
+
+    @SuppressWarnings("unused")
+    public String getUserMail() {
+        return userMail;
+    }
 }
